refactor Logger class and move instance configuration to main
The Logger class has been refactored to be more lightweight and only handle basic functionality, removing the creation and configuration of logging sinks. These functionalities are moved to the main.cpp file. This changes allows more flexibility in configuring the Logger instance and reduces the responsibility of the Logger class.